Comparing ppt, videos, excel file added
