+++ compare_jpg.py
