+++ compare_ppt.py
